Splitted specs + moved before/afterEach to global onPrepare (#4)
